refactor: 💡 extract functions to MatterJSUtils
