+++ src/components/DataRoom/USDCStorage/MatterCanvas.tsx
@@ -1,24 +1,14 @@
 import React, { useCallback, useEffect, useRef, useState } from 'react'
 import Matter from 'matter-js'
 import { debounce } from 'lodash'
+import type { MotionValue } from 'framer-motion'
+import { addWallsToWorld, createCoin, createWalls, type Dimensions } from './MatterJSUtils'
 
-const { Bodies, Engine, Render, Runner, World } = Matter
-
-const WALL_BORDER_WIDTH = 25
-const WALL_LENGTH = 500
-const COIN_RADIUS = 50
-
-const IMG_TEXTURE_SIZE = 256 // Size Of The USDC.png Image In Pixels
-const IMG_SCALE = (COIN_RADIUS * 2) / IMG_TEXTURE_SIZE // Scale Factor To Match Image Texture Size With Coin Radius
+const { Engine, Render, Runner, World } = Matter
 
 type MatterCanvasProps = {
   imgUrl: string
-  scrollYProgress: any // Specify the correct type based on your usage
-}
-
-type Dimensions = {
-  width: number
-  height: number
+  scrollYProgress: MotionValue<number>
 }
 
 export default function MatterCanvas({ imgUrl, scrollYProgress }: MatterCanvasProps) {
@@ -42,43 +32,6 @@ export default function MatterCanvas({ imgUrl, scrollYProgress }: MatterCanvasPr
         height: dimensions.height,
         background: '#121312',
         wireframes: false,
-      },
-    })
-  }
-
-  const createWalls = (dimensions: Dimensions) => [
-    { x: 0, y: -WALL_BORDER_WIDTH / 2, width: WALL_LENGTH * 8, height: WALL_BORDER_WIDTH },
-    {
-      x: 0,
-      y: dimensions.height + WALL_BORDER_WIDTH / 2 - 15,
-      width: dimensions.width * 9,
-      height: WALL_BORDER_WIDTH,
-    },
-    { x: 0, y: dimensions.height / 2, width: WALL_BORDER_WIDTH, height: WALL_LENGTH * 8 },
-    { x: dimensions.width, y: dimensions.height / 2, width: WALL_BORDER_WIDTH, height: WALL_LENGTH * 8 },
-  ]
-
-  const addWallsToWorld = (engine: Matter.Engine, walls: any[]) => {
-    World.add(
-      engine.world,
-      walls.map((wall) =>
-        Bodies.rectangle(wall.x, wall.y, wall.width, wall.height, {
-          isStatic: true,
-          render: { visible: false },
-        }),
-      ),
-    )
-  }
-
-  const createCoin = (dimensions: Dimensions, imgUrl: string) => {
-    return Bodies.circle(Math.random() * dimensions.width * 0.5 + dimensions.width * 0.25, 0, COIN_RADIUS, {
-      restitution: 0.4,
-      render: {
-        sprite: {
-          texture: imgUrl,
-          xScale: IMG_SCALE,
-          yScale: IMG_SCALE,
-        },
       },
     })
   }
